[F61iNGQX] Splits the integration tests into a separate package
[F61iNGQX] Makes the test run in parallel

[F61iNGQX] Tries to run projects in parallel

Would printing less spead up the builds?

Tries to run core in parallel

Ignores AtomicTest

Uses temurin as JVM and decreases the heap size

Tries overriding the gradle.properties

Source
https://github.com/actions/runner-images/issues/6709
https://github.com/gradle/gradle/issues/19750

Tries to fix number of threads to 4

Adds separate ci step

Adds separate step in CI

Downloads docker containers
