+++ common/build.gradle
@@ -29,6 +29,10 @@ generateGrammarSource {
     arguments += ["-package", "apoc.custom"]
 }
 
+test {
+    maxParallelForks = Runtime.runtime.availableProcessors().intdiv(2) ?: 1
+}
+
 dependencies {
     apt project(':processor')
     apt group: 'org.neo4j', name: 'neo4j', version: neo4jVersionEffective
