+++ core/build.gradle
@@ -45,6 +45,10 @@ generateGrammarSource {
     arguments += ["-package", "apoc.custom"]
 }
 
+test {
+    maxParallelForks = Runtime.runtime.availableProcessors().intdiv(2) ?: 1
+}
+
 dependencies {
     apt project(':processor')
     apt group: 'org.neo4j', name: 'neo4j', version: neo4jVersionEffective
@@ -92,7 +96,6 @@ dependencies {
         exclude group: 'org.slf4j', module: 'slf4j-api'
     }
     testImplementation group: 'com.github.adejanovski', name: 'cassandra-jdbc-wrapper', version: '3.1.0'
-    testImplementation group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-csv', version: '2.13.2'
 
     configurations.all {
         exclude group: 'org.slf4j', module: 'slf4j-nop'
