+++ gradle.properties
@@ -1,0 +1,2 @@
+org.gradle.parallel=true
+rg.gradle.jvmargs=-Xmx4g -XX:MaxMetaspaceSize=256m
