+++ it/src/test/java/apoc/it/core/ApocSplitTest.java
@@ -1,4 +1,4 @@
-package apoc;
+package apoc.it.core;
 
 import apoc.util.Neo4jContainerExtension;
 import apoc.util.TestContainerUtil;
