+++ it/src/test/java/apoc/it/core/ExportCypherEnterpriseFeaturesTest.java
@@ -1,5 +1,4 @@
-package apoc.export.cypher;
-
+package apoc.it.core;
 import apoc.util.Neo4jContainerExtension;
 import apoc.util.TestContainerUtil.ApocPackage;
 import apoc.util.TestUtil;
