+++ it/src/test/java/apoc/it/core/GraphRefactoringEnterpriseTest.java
@@ -1,4 +1,4 @@
-package apoc.refactor;
+package apoc.it.core;
 
 import apoc.util.Neo4jContainerExtension;
 import apoc.util.TestContainerUtil.ApocPackage;
