+++ processor/build.gradle
@@ -6,6 +6,9 @@ plugins {
 
 archivesBaseName = "apoc-processor"
 
+test {
+    maxParallelForks = Runtime.runtime.availableProcessors().intdiv(2) ?: 1
+}
 
 jar {
     manifest {
