+++ settings.gradle
@@ -4,4 +4,4 @@ include("processor")
 include("common")
 include("core")
 include("test-utils")
-include("test-startup")
+include("it")
