+++ test-utils/build.gradle
@@ -11,6 +11,10 @@ javadoc {
     options.addStringOption('Xdoclint:none', '-quiet')
 }
 
+test {
+    maxParallelForks = Runtime.runtime.availableProcessors().intdiv(2) ?: 1
+}
+
 dependencies {
     def withoutServers = {
         exclude group: 'org.eclipse.jetty'
@@ -51,6 +55,7 @@ dependencies {
     implementation group: 'org.gradle', name: 'gradle-tooling-api', version: '7.2'
     implementation group: 'org.jetbrains', name: 'annotations', version: "17.0.0"
     implementation group: 'com.amazonaws', name: 'aws-java-sdk-s3', version: '1.12.348'
+    implementation group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-csv', version: '2.13.2'
 }
