+++ test-utils/src/main/java/apoc/csv/CsvTestUtil.java
@@ -1,4 +1,4 @@
-package apoc.export.csv;
+package apoc.csv;
 
 import com.fasterxml.jackson.databind.MappingIterator;
 import com.fasterxml.jackson.dataformat.csv.CsvMapper;
